Merge

+++ dataBaseFirebird/models/modeluser.cpp
@@ -1,6 +1,7 @@
 #include "modeluser.h"
 #include <QSqlQuery>
 #include <QSqlError>
+#include <QDate>
 #include <QDebug>
 
 ModelUser::ModelUser(QObject *parent) : QObject(parent)
@@ -9,23 +10,35 @@ ModelUser::ModelUser(QObject *parent) : QObject(parent)
 
 ModelUser::ModelUser(QSqlDatabase db, QObject *parent) : QObject(parent) {
     data_base = db;
+    mBx = new QMessageBox();
 }
 
 void ModelUser::addUser() {
     addUserDialog * add_user = new addUserDialog();
     add_user->show();
-    connect(add_user, SIGNAL(onOkClick(int,int,QString,QString,int)), SLOT(finishAddUser(int,int,QString,QString,int)));
+    connect(add_user, SIGNAL(onOkClick(int,QString,QString, int, QString)), SLOT(finishAddUser(int,QString,QString, int, QString)));
 }
 
 void ModelUser::deleteUsers(int user_id) {
+    QSqlQuery* query0 = new QSqlQuery(data_base);
+    QString statament0 = QString("SELECT COUNT(userid) FROM users WHERE parentid=%1").arg(QString::number(user_id));
+    query0->exec(statament0);
+    query0->next();
+    if (query0->value(0).toInt() > 0) {
+        mBx->setWindowTitle("Удаление пользователя");
+        mBx->setText("Невозможно удалить пользователя с зависимостями");
+        mBx->show();
+        return;
+    }
+
     QSqlQuery* query = new QSqlQuery(data_base);
-    QString statament = "DELETE FROM users WHERE userid=";
-    statament.append(QString::number(user_id));
+    QString statament = QString("DELETE FROM users WHERE userid=%1").arg(QString::number(user_id));
 
     query->exec(statament);
     if (query->lastError().isValid()) {
-        //QMessage
-        qDebug() << "RemoveRow:" << query->lastError();
+        mBx->setWindowTitle("Удаление пользователя");
+        mBx->setText(query->lastError().databaseText());
+        mBx->show();
     }
     emit needUpdate();
 }
@@ -33,41 +46,54 @@ void ModelUser::deleteUsers(int user_id) {
 void ModelUser::editUsers(int user_id) {
     addUserDialog * add_user = new addUserDialog(user_id, data_base);
     add_user->show();
-    connect(add_user, SIGNAL(onOkClick(int, int,int,QString,QString,int)), SLOT(finishEditUser(int,int,int,QString,QString,int)));
+    connect(add_user, SIGNAL(onOkClick(int,int,QString,QString, int, QString, bool)), SLOT(finishEditUser(int,int,QString,QString, int, QString, bool)));
 }
 
-void ModelUser::finishAddUser(int parentId, int cardId, QString name, QString shortName, int indexId) {
+void ModelUser::finishAddUser(int parentId, QString name, QString shortName, int perm, QString date) {
+    QString dt = QDate::currentDate().toString("dd.MM.yy");
+    if (!date.isEmpty()) {
+        dt = date;
+    }
+
     QSqlQuery* query = new QSqlQuery(data_base);
-    QString statament = "INSERT INTO users (parentId, shortname, viewname, cardid, indexid) VALUES (";
-    statament.append(QString::number(parentId) + ", ");
-    statament.append("'" + shortName + "', ");
-    statament.append("'" + name + "', ");
-    statament.append(QString::number(cardId) + ", ");
-    statament.append(QString::number(indexId));
-    statament.append(")");
+    QString statament = QString("INSERT INTO users (parentId, shortname, viewname, cardid, flags, sldate, refslim) VALUES (%1, %2, %3, %4, %5, '%6', 1)")
+            .arg(QString::number(parentId))
+            .arg("'" + shortName + "'")
+            .arg("'" + name + "'")
+            .arg("0")
+            .arg(QString::number(perm))
+            .arg(dt);
 
     query->exec(statament);
     if (query->lastError().isValid()) {
-        //добавить QMessage
-        qDebug() << "AddUser:" << query->lastError();
+        qDebug() << query->lastError().databaseText();
+        mBx->setWindowTitle("Добавление пользователя");
+        mBx->setText(query->lastError().databaseText());
+        mBx->show();
     }
     emit needUpdate();
 }
 
-void ModelUser::finishEditUser(int userId, int parentId, int cardId, QString name, QString shortName, int indexId) {
+void ModelUser::finishEditUser(int userId, int parentId, QString name, QString shortName, int perm, QString date, bool dateUpd) {
+    //date and dt compate
+    int dateUpdated = dateUpd ? 0 : 1;
+
     QSqlQuery* query = new QSqlQuery(data_base);
-    QString statament = QString("UPDATE users SET parentid=%1, shortname=%2, viewname=%3, cardid=%4, indexid=%5 WHERE userid=%6")
+    QString statament = QString("UPDATE users SET parentid=%1, shortname=%2, viewname=%3, flags=%4, sldate='%5', refslim=%6 WHERE userid=%7")
             .arg(QString::number(parentId))
             .arg("'" +  shortName + "'")
             .arg("'" + name + "'")
-            .arg(QString::number(cardId))
-            .arg(QString::number(indexId))
+            .arg(QString::number(perm))
+            .arg(date)
+            .arg(dateUpdated)
             .arg(QString::number(userId));
 
     query->exec(statament);
     if (query->lastError().isValid()) {
-        //QMessage
-        qDebug() << "Error:" << query->lastError();
+        qDebug() << query->lastError().databaseText();
+        mBx->setWindowTitle("Редактирование пользователя");
+        mBx->setText(query->lastError().databaseText());
+        mBx->show();
     }
     emit needUpdate();
 }
